test: reorganize and deduplicate service tests
