+++ packages/core/src/services/authz-io/__tests__/authz-io-local.service.spec.ts
@@ -14,49 +14,214 @@
  * limitations under the License.
  */
 
-import { UnitObject } from '@univerjs/protocol';
-import { beforeEach, describe, expect, it } from 'vitest';
-import { Injector } from '../../../common/di';
-import { DesktopLogService, ILogService } from '../../log/log.service';
-import { ResourceManagerService } from '../../resource-manager/resource-manager.service';
-import { IResourceManagerService } from '../../resource-manager/type';
+import type { Univer } from '../../../univer';
+import { UnitAction, UnitObject, UnitRole } from '@univerjs/protocol';
+import { BehaviorSubject } from 'rxjs';
+import { afterEach, beforeEach, describe, expect, it } from 'vitest';
+import { UnitModel, UniverInstanceType } from '../../../common/unit';
+import { createTestBed } from '../../__tests__/create-test-bed';
+import { IUniverInstanceService } from '../../instance/instance.service';
+import { IResourceLoaderService } from '../../resource-loader/type';
 import { UserManagerService } from '../../user-manager/user-manager.service';
-import { AuthzIoLocalService } from '../authz-io-local.service';
+import { IAuthzIoService } from '../type';
+
+interface ITestBoardData {
+    id: string;
+    name?: string;
+    resources?: Array<{ name: string; data: string }>;
+}
+
+class MockBoardUnit extends UnitModel<ITestBoardData, UniverInstanceType.UNIVER_BOARD> {
+    override readonly type = UniverInstanceType.UNIVER_BOARD;
+    override name$ = new BehaviorSubject('');
+    private readonly _snapshot: ITestBoardData;
+
+    constructor(snapshot: Partial<ITestBoardData> = {}) {
+        super();
+        this._snapshot = {
+            id: 'board-resource',
+            name: '',
+            ...snapshot,
+        };
+        this.name$.next(this._snapshot.name ?? '');
+    }
+
+    override getUnitId(): string {
+        return this._snapshot.id;
+    }
+
+    override setName(name: string): void {
+        this._snapshot.name = name;
+        this.name$.next(name);
+    }
+
+    override getSnapshot(): ITestBoardData {
+        return this._snapshot;
+    }
+
+    override getRev(): number {
+        return 1;
+    }
+
+    override incrementRev(): void { }
+
+    override setRev(): void { }
+}
 
 describe('AuthzIoLocalService', () => {
-    let service: AuthzIoLocalService;
+    let univer: Univer;
+    let unitId: string;
 
     beforeEach(() => {
-        const injector = new Injector();
-        injector.add([ILogService, { useClass: DesktopLogService }]);
-        injector.add([IResourceManagerService, { useClass: ResourceManagerService }]);
-        injector.add([UserManagerService]);
-        injector.add([AuthzIoLocalService]);
-        service = injector.get(AuthzIoLocalService);
+        const instance = createTestBed();
+        univer = instance.univer;
+        unitId = instance.unitId;
     });
 
-    it('creates range permission objects and evaluates default owner actions', async () => {
-        const objectID = await service.create({
-            objectType: UnitObject.SelectRange,
-            selectRangeObject: { unitID: 'book-1', name: 'Protected range' },
-        } as never);
+    afterEach(() => {
+        univer.dispose();
+    });
 
-        await expect(service.allowed({ unitID: 'book-1', objectID, actions: [6, 16] } as never)).resolves.toEqual([
-            { action: 6, allowed: true },
-            { action: 16, allowed: true },
+    it('should persist and reload permission resources through a real workbook lifecycle', async () => {
+        const injector = univer.__getInjector();
+        const authzIoService = injector.get(IAuthzIoService);
+        const resourceLoaderService = injector.get(IResourceLoaderService);
+        const userManagerService = injector.get(UserManagerService);
+
+        userManagerService.setCurrentUser({
+            userID: 'Owner_real-user',
+            name: 'Owner User',
+            avatar: '',
+        });
+
+        const objectID = await authzIoService.create({
+            objectType: UnitObject.SelectRange,
+            selectRangeObject: {
+                unitID: unitId,
+                name: 'Protected range',
+                collaborators: [],
+                scope: undefined,
+            },
+        });
+
+        expect(await authzIoService.allowed({ unitID: unitId, objectID, objectType: UnitObject.SelectRange, actions: [UnitAction.Edit] })).toEqual([
+            { action: UnitAction.Edit, allowed: true },
         ]);
+
+        await authzIoService.update({
+            objectType: UnitObject.SelectRange,
+            objectID,
+            unitID: unitId,
+            share: undefined,
+            name: 'Protected range',
+            strategies: [{ action: UnitAction.Edit, role: UnitRole.Reader }],
+            scope: undefined,
+            collaborators: undefined,
+        });
+
+        expect(await authzIoService.allowed({ unitID: unitId, objectID, objectType: UnitObject.SelectRange, actions: [UnitAction.Edit] })).toEqual([
+            { action: UnitAction.Edit, allowed: false },
+        ]);
+
+        const listed = await authzIoService.list({
+            unitID: unitId,
+            objectIDs: [objectID],
+            actions: [UnitAction.Edit, UnitAction.View],
+        });
+
+        expect(listed).toHaveLength(1);
+        expect(listed[0].name).toBe('Protected range');
+        expect(listed[0].actions).toContainEqual({ action: UnitAction.Edit, allowed: false });
+
+        const batched = await authzIoService.batchAllowed([
+            { unitID: unitId, objectID, objectType: UnitObject.SelectRange, actions: [UnitAction.Edit] },
+            { unitID: unitId, objectID: 'missing', objectType: UnitObject.SelectRange, actions: [UnitAction.View] },
+        ]);
+
+        expect(batched[0].actions).toEqual([{ action: UnitAction.Edit, allowed: false }]);
+        expect(batched[1].actions).toEqual([{ action: UnitAction.View, allowed: true }]);
+
+        expect(await authzIoService.listCollaborators({ unitID: unitId, objectID })).toEqual([]);
+        expect(await authzIoService.listRoles({ objectType: UnitObject.SelectRange })).toEqual({ roles: [], actions: [] });
+        await expect(authzIoService.createCollaborator({ unitID: unitId, objectID } as never)).resolves.toBeUndefined();
+        await expect(authzIoService.updateCollaborator({ unitID: unitId, objectID } as never)).resolves.toBeUndefined();
+        await expect(authzIoService.deleteCollaborator({ unitID: unitId, objectID } as never)).resolves.toBeUndefined();
+        await expect(authzIoService.putCollaborators({ unitID: unitId, objectID, collaborators: [] })).resolves.toBeUndefined();
+
+        const snapshot = resourceLoaderService.saveUnit(unitId);
+        const authzResource = snapshot?.resources.find((resource) => resource.name === 'SHEET_AuthzIoMockService_PLUGIN');
+
+        expect(authzResource?.data).toContain(objectID);
+
+        injector.get(IUniverInstanceService).disposeUnit(unitId);
+
+        const unloaded = await authzIoService.list({
+            unitID: unitId,
+            objectIDs: [objectID],
+            actions: [UnitAction.Edit],
+        });
+
+        expect(unloaded[0].name).toBe('');
+        expect(unloaded[0].actions).toEqual([{ action: UnitAction.Edit, allowed: false }]);
+
+        univer.createUnit(UniverInstanceType.UNIVER_SHEET, snapshot!);
+
+        const reloaded = await authzIoService.list({
+            unitID: unitId,
+            objectIDs: [objectID],
+            actions: [UnitAction.Edit],
+        });
+
+        expect(reloaded[0].name).toBe('Protected range');
+        expect(reloaded[0].actions).toEqual([{ action: UnitAction.Edit, allowed: false }]);
     });
 
-    it('lists created permission objects with their range metadata', async () => {
-        const objectID = await service.create({
-            objectType: UnitObject.SelectRange,
-            selectRangeObject: { unitID: 'book-1', name: 'Protected range' },
-        } as never);
+    it('should persist and reload permission resources for board units', async () => {
+        const injector = univer.__getInjector();
+        const authzIoService = injector.get(IAuthzIoService);
+        const resourceLoaderService = injector.get(IResourceLoaderService);
+        const univerInstanceService = injector.get(IUniverInstanceService);
 
-        const [permissionPoint] = await service.list({ unitID: 'book-1', objectIDs: [objectID], actions: [6] } as never);
+        univerInstanceService.registerCtorForType(UniverInstanceType.UNIVER_BOARD, MockBoardUnit);
+        const board = univer.createUnit<ITestBoardData, MockBoardUnit>(UniverInstanceType.UNIVER_BOARD, {
+            id: 'board-resource',
+        });
+        const objectID = await authzIoService.create({
+            objectType: UnitObject.Workbook,
+            worksheetObject: {
+                unitID: board.getUnitId(),
+                name: 'Board permission',
+                collaborators: [],
+                strategies: [{ action: UnitAction.Edit, role: UnitRole.Owner }],
+                scope: undefined,
+            },
+        });
 
-        expect(permissionPoint.objectID).toBe(objectID);
-        expect(permissionPoint.name).toBe('Protected range');
-        expect(permissionPoint.unitID).toBe('book-1');
+        const snapshot = resourceLoaderService.saveUnit<ITestBoardData>(board.getUnitId());
+        const authzResource = snapshot?.resources?.find((resource) => resource.name === 'SHEET_AuthzIoMockService_PLUGIN');
+
+        expect(authzResource?.data).toContain(objectID);
+
+        expect(univerInstanceService.disposeUnit(board.getUnitId())).toBe(true);
+
+        const unloaded = await authzIoService.list({
+            unitID: board.getUnitId(),
+            objectIDs: [objectID],
+            actions: [UnitAction.Edit],
+        });
+
+        expect(unloaded[0].name).toBe('');
+        expect(unloaded[0].actions).toEqual([{ action: UnitAction.Edit, allowed: true }]);
+
+        univer.createUnit<ITestBoardData, MockBoardUnit>(UniverInstanceType.UNIVER_BOARD, snapshot!);
+
+        const reloaded = await authzIoService.list({
+            unitID: board.getUnitId(),
+            objectIDs: [objectID],
+            actions: [UnitAction.Edit],
+        });
+
+        expect(reloaded[0].name).toBe('Board permission');
+        expect(reloaded[0].actions).toEqual([{ action: UnitAction.Edit, allowed: true }]);
     });
 });
