+++ packages/sheets-sort/src/commands/commands/__tests__/create-command-test-bed.ts
@@ -15,7 +15,19 @@
  */
 
 import type { Dependency, IWorkbookData, Workbook } from '@univerjs/core';
-import { BooleanNumber, ILogService, Inject, Injector, IUniverInstanceService, LocaleType, LogLevel, Plugin, Tools, Univer, UniverInstanceType } from '@univerjs/core';
+import {
+    BooleanNumber,
+    ILogService,
+    Inject,
+    Injector,
+    IUniverInstanceService,
+    LocaleType,
+    LogLevel,
+    Plugin,
+    Tools,
+    Univer,
+    UniverInstanceType,
+} from '@univerjs/core';
 import { FormulaDataModel } from '@univerjs/engine-formula';
 import { SheetInterceptorService, SheetSkeletonService, SheetsSelectionsService } from '@univerjs/sheets';
 import { SheetsSortController } from '../../../controllers/sheets-sort.controller';
