+++ packages/sheets-ui/src/services/utils/defined-name-utils.ts
@@ -15,7 +15,13 @@
  */
 
 import type { IUniverInstanceService, Workbook } from '@univerjs/core';
-import type { IDefinedNamesService, IDefinedNamesServiceParam, IFunctionService, ISuperTableService, LexerTreeBuilder } from '@univerjs/engine-formula';
+import type {
+    IDefinedNamesService,
+    IDefinedNamesServiceParam,
+    IFunctionService,
+    ISuperTableService,
+    LexerTreeBuilder,
+} from '@univerjs/engine-formula';
 import type { ISelectionWithStyle } from '@univerjs/sheets';
 import { AbsoluteRefType } from '@univerjs/core';
 import { isReferenceStringWithEffectiveColumn, serializeRangeWithSheet } from '@univerjs/engine-formula';
