+++ packages/sheets-ui/src/services/utils/doc-skeleton-util.ts
@@ -15,8 +15,20 @@
  */
 
 import type { ICellWithCoord, ICustomRange, Injector, IParagraph, ITextRangeParam, Workbook } from '@univerjs/core';
-import type { DocumentSkeleton, IBoundRectNoAngle, IDocumentSkeletonGlyph, IFontCacheItem } from '@univerjs/engine-render';
-import { CustomRangeType, HorizontalAlign, IUniverInstanceService, PresetListType, UniverInstanceType, VerticalAlign } from '@univerjs/core';
+import type {
+    DocumentSkeleton,
+    IBoundRectNoAngle,
+    IDocumentSkeletonGlyph,
+    IFontCacheItem,
+} from '@univerjs/engine-render';
+import {
+    CustomRangeType,
+    HorizontalAlign,
+    IUniverInstanceService,
+    PresetListType,
+    UniverInstanceType,
+    VerticalAlign,
+} from '@univerjs/core';
 import { DocSkeletonManagerService } from '@univerjs/docs';
 import { DOC_VERTICAL_PADDING, getLineBounding, NodePositionConvertToCursor } from '@univerjs/docs-ui';
 import { IRenderManagerService } from '@univerjs/engine-render';
